Split the responsibilities of playermovement
